apr.h: __attribute__ and __has_attribute exist for !__GNUC__ compilers.  PR 69009.
There are more compilers than __GNUC__ that provide __attribute__ so let's not
no-op it based on (un)known compilers but simply if it's not already defined
(it should be provided as a pre-defined macro if defined by the compiler).

While at, do the same for __has_attribute.

Submitted by: Petr Sumbera <petr.sumbera oracle.com>



git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1917741 13f79535-47bb-0310-9956-ffa450edef68